+++ AnimMeteor.h
@@ -6,17 +6,17 @@
 
 class AnimMeteor{
   public:
-    AnimMeteor(CRGB *ledArray, int stripSize);
+    AnimMeteor(CRGB *ledArray, uint8_t stripSize);
     void Animate(CRGB color, byte meteorSize, byte meteorTrailDecay, boolean meteorRandomDecay);
     void AnimateInverted(CRGB color, byte meteorSize, byte meteorTrailDecay, boolean meteorRandomDecay);
     void Setup(bool isInverted);
   private:
-    void FadeToBlack(int ledNo, byte fadeValue);
+    void FadeToBlack(uint8_t ledNo, byte fadeValue);
     CRGB *_ledArray;
-    int _animationStep;
-    int _ledStripSize;
-    int _animationStartIndex;
-    int _animationEndIndex;
+    uint8_t _animationStep;
+    uint8_t _ledStripSize;
+    uint8_t _animationStartIndex;
+    uint8_t _animationEndIndex;
     bool _isInverted;
 };
