+++ LeftToRightIndexer.cpp
@@ -9,24 +9,23 @@ void LeftToRightIndexer::Setup(CRGB *left, CRGB *right, CRGB *bottom){//, CRGB *
   //_top = top;
 }
 
-void LeftToRightIndexer::SetColor(int index, CRGB color){
+void LeftToRightIndexer::SetColor(uint8_t index, CRGB color){
   if(index < NUM_LEDS_VERTICAL){
-    int idx = NUM_LEDS_VERTICAL-index-1;
+    uint8_t idx = NUM_LEDS_VERTICAL-index-1;
     _left[idx] = color;
     DBG("L\n");    
   }
 
   else if(index >= NUM_LEDS_VERTICAL && index < NUM_LEDS_VH){
-    int idx = NUM_LEDS_VH-1-index;
+    uint8_t idx = NUM_LEDS_VH-1-index;
     _bottom[idx] = color;
     DBG("B\n");
   }
 
   else{
-    int idx = index - NUM_LEDS_VH;
+    uint8_t idx = index - NUM_LEDS_VH;
     // DBG("index:%u VH:%u idx:%u\n", index, NUM_LEDS_VH, idx);
     _right[idx] = color;
     DBG("R\n");
   }
 }
-
