+++ LeftToRightIndexer.h
@@ -8,7 +8,7 @@ class LeftToRightIndexer{
   public:
     void Setup(CRGB *left, CRGB *right, CRGB *bottom);
     //LeftToRightIndexer(CRGB *left, CRGB *right, CRGB *bottom, CRGB *top=NULL);
-    void SetColor(int index, CRGB color);
+    void SetColor(uint8_t index, CRGB color);
   private:
     CRGB *_left;
     CRGB *_right;
