+++ SymVerticalIndexer.h
@@ -8,8 +8,8 @@ class SymVerticalIndexer{
   public:
     void Setup(CRGB *left, CRGB *right, CRGB *bottom);
     //LeftToRightIndexer(CRGB *left, CRGB *right, CRGB *bottom, CRGB *top=NULL);
-    void SetColor(int index, CRGB color);
-    CRGB GetColor(int index);
+    void SetColor(uint8_t index, CRGB color);
+    CRGB GetColor(uint8_t index);
   private:
     CRGB *_left;
     CRGB *_right;
